@@ -7,6 +7,6 @@
     <PackageReference Include="CloudinaryDotNet" Version="1.11.0"/>
   </ItemGroup>
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 </Project>
