@@ -1,6 +1,7 @@
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
 using Domain;
-using Newtonsoft.Json;
+
 
 namespace Application.Profiles
 {
@@ -11,7 +12,7 @@ namespace Application.Profiles
         public string Image { get; set; }
         public string Bio { get; set; }
 
-        [JsonProperty("following")]
+        [JsonPropertyName("following")]
         public bool IsFollowed { get; set; }
         public int FollowersCount { get; set; }
         public int FollowingCount { get; set; }
